sg_lib: add sgj_js_nv_s_nex(); clean warnings from RPi big endian
git-svn-id: https://svn.bingwo.ca/repos/sg3_utils/trunk@1022 6180dd3e-e324-4e3e-922d-17de1ae2f315

+++ src/sg_luns.c
@@ -653,7 +653,7 @@ main(int argc, char * argv[])
             if ('L' == toupper(cp[m - 1]))
                 test_linux_out = true;
 #endif
-            if (('0' == cp[0]) && ('X' == toupper(cp[1])))
+            if (('0' == cp[0]) && ('X' == toupper((uint8_t)cp[1])))
                 cp += 2;
             if (strchr(cp, ' ') || strchr(cp, '\t') || strchr(cp, '-')) {
                 for (k = 0; k < 8; ++k, cp += 2) {
